# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Swarm API",
    "version": "1.74.1",
    "description": "Multi-agent orchestration API for Claude Code, Codex, and Gemini CLI. Enables task distribution, agent communication, and service discovery.\n\nMCP tools are documented separately in [MCP.md](./MCP.md)."
  },
  "servers": [
    {
      "url": "http://localhost:3013",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key via Authorization: Bearer <API_KEY>"
      },
      "agentId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-ID",
        "description": "Agent UUID for agent-scoped operations"
      }
    },
...
  },
  "paths": {
...
    "/api/workflows/{id}/nodes/{nodeId}": {
      "patch": {
        "summary": "Patch a single node in a workflow definition",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "nodeId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "description": "Executor type: 'agent-task', 'script', 'raw-llm', 'validate', 'property-match'"
                  },
                  "label": {
                    "type": "string",
                    "description": "Human-readable label for UI display"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {},
                    "description": "Executor-specific config. For agent-task: { template, outputSchema?, agentId?, tags?, priority?, dir?, vcsRepo?, model? }. Values support {{interpolation}} from the node's inputs context. NOTE: config.outputSchema on agent-task nodes validates the AGENT's raw JSON output, while node-level outputSchema validates the EXECUTOR's return value ({taskId, taskOutput})."
                  },
                  "next": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      {
                        "type": "object",
                        "additionalProperties": {
                          "type": "string"
                        }
                      }
                    ],
                    "description": "Next node(s): string for simple chaining, string[] for fan-out to parallel nodes, or record for port-based routing ({pass: 'a', fail: 'b'})"
                  },
                  "validation": {
                    "type": "object",
                    "properties": {
                      "executor": {
                        "type": "string",
                        "default": "validate"
                      },
                      "config": {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      "mustPass": {
                        "type": "boolean",
                        "default": false
                      },
                      "retry": {
                        "type": "object",
                        "properties": {
                          "maxRetries": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 3
                          },
                          "strategy": {
                            "type": "string",
                            "enum": [
                              "exponential",
                              "static",
                              "linear"
                            ],
                            "default": "exponential"
                          },
                          "baseDelayMs": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 1000
                          },
                          "maxDelayMs": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 60000
                          }
                        }
                      }
                    },
                    "required": [
                      "config"
                    ]
                  },
                  "retry": {
                    "type": "object",
                    "properties": {
                      "maxRetries": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 3
                      },
                      "strategy": {
                        "type": "string",
                        "enum": [
                          "exponential",
                          "static",
                          "linear"
                        ],
                        "default": "exponential"
                      },
                      "baseDelayMs": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 1000
                      },
                      "maxDelayMs": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 60000
                      }
                    }
                  },
                  "inputs": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "REQUIRED for cross-node data access. Maps local names to context paths. Without this, upstream step outputs are NOT available for interpolation — only 'trigger' and 'input' are. Example: { \"cityData\": \"generate-city\" } → use {{cityData.taskOutput.field}} in config templates. For trigger data: { \"pr\": \"trigger.pullRequest\" }."
                  },
                  "inputSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "description": "JSON Schema to validate resolved inputs before execution"
                  },
                  "outputSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "description": "JSON Schema to validate the executor's output (e.g. {taskId, taskOutput} for agent-task). Different from config.outputSchema which validates the agent's raw output."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node patched (version snapshot created)"
          },
          "400": {
            "description": "Invalid patch or resulting definition"
          },
          "404": {
            "description": "Workflow or node not found"
          }
        }
      }
    },
    "/api/workflows/{id}/trigger": {
      "post": {
        "summary": "Trigger a workflow execution",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Workflow run started (or skipped if cooldown active)"
          },
          "400": {
            "description": "Workflow is disabled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      }
    },
    "/api/workflows/{id}/runs": {
      "get": {
        "summary": "List runs for a workflow",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "running",
                "waiting",
                "completed",
                "failed",
                "skipped",
                "cancelled"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow run list"
          }
        }
      }
    },
    "/api/workflow-runs/{id}": {
      "get": {
        "summary": "Get a workflow run with steps (includes retry columns)",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow run details with steps including retry info"
          },
          "404": {
            "description": "Run not found"
          }
        }
      }
    },
    "/api/workflow-runs/{id}/retry": {
      "post": {
        "summary": "Retry a failed workflow run",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Retry started"
          },
          "400": {
            "description": "Cannot retry"
          }
        }
      }
    },
    "/api/workflow-runs/{id}/cancel": {
      "post": {
        "summary": "Cancel a running or waiting workflow run",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run cancelled"
          },
          "400": {
            "description": "Cannot cancel"
          }
        }
      }
    },
    "/api/executor-types": {
      "get": {
        "summary": "List all executor types with their config and output schemas",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of executor types with schemas"
          }
        }
      }
    },
    "/api/executor-types/{type}": {
      "get": {
        "summary": "Get a specific executor type with its schemas",
        "tags": [
          "Workflows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "type",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Executor type details"
          },
          "404": {
            "description": "Executor type not found"
          }
        }
      }
    },
    "/api/webhooks/{workflowId}": {
      "post": {
        "summary": "Trigger workflow via webhook",
        "tags": [
          "Webhooks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workflowId",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Webhook processed"
          },
          "401": {
            "description": "Invalid signature"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      }
    },
...
  },
...
}